server/requirements.txt
face_recognition_models
Click>=6.0
numpy
Pillow
scipy>=0.17.0
opencv-contrib-python==4.5.5.64
opencv-python==4.5.5.64
face_recognition
django
imutils
mediapipe
art==5.6
loguru
gunicorn
deskew